+++ doc/users/doxygen.html
@@ -30,6 +30,7 @@
       Doxygen, if the online documentation is available,
       <a href="../doxygen/deal.II/index.html">here</a>. For older revisions,
 	links are provided here by revision number:
+	<a href="http://www.dealii.org/8.3.0/doxygen/deal.tag">8.3.0</a>,
 	<a href="http://www.dealii.org/8.2.0/doxygen/deal.tag">8.2.0</a>,
 	<a href="http://www.dealii.org/8.1.0/doxygen/deal.tag">8.1.0</a>,
 	<a href="http://www.dealii.org/8.0.0/doxygen/deal.tag">8.0.0</a>,
